requirements.txt
# PyTorch and Torchvision installed from source for Jetson-Nano
# Please comment torch & torchvision when working on Jetson-Nano

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.1
flask
numpy
Pillow
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
imageio
argparse
imageio
scipy
pyyaml
matplotlib
wandb
